+++ cranelift/assembler-x64/Cargo.toml
@@ -8,9 +8,12 @@ rust-version.workspace = true
 
 [dependencies]
 arbitrary = { workspace = true, features = ["derive"], optional = true }
+capstone = { workspace = true, optional = true }
 
 [dev-dependencies]
+arbitrary = { workspace = true, features = ["derive"] }
 arbtest = "0.3.1"
+capstone = { workspace = true }
 
 [build-dependencies]
 cranelift-assembler-x64-meta = { path = "meta", version = "0.117.0" }
@@ -23,4 +26,4 @@ similar_names = { level = "allow", priority = 1 }
 wildcard_imports = { level = "allow", priority = 1 }
 
 [features]
-arbitrary = ['dep:arbitrary']
+fuzz = ['dep:arbitrary', 'dep:capstone']
